Add functionality for buttons
The functionality behind the buttons 'Clear All' and
'Check failed' has been enabled. As part of this I've
changed the comparison in `CheckFailedNodesVisitor.Visit`
as otherwise reference equality is used.

I've tried to follow the general NUnit code guidelines
which makes the diff look strange due to mix of spaces
and tabs in the existing code.

Fixes #28